Added package for 'libaec'.
